Update guides.md
Please first approve my earlier PR for the update to bestpractices.md. 
This PR adds a sentence under the Simulation section that references the new section in bestpractices.md. 
I'm not confident in how I linked to the reference in best practices, so please double-check that. 
I wrote:
We recommend reading the Best Practices Guide for [Simulation Data](../../bestpractices/working#simulationdata) as a prerequesite to initiating your project.